Explore ways of specifying immediately committed with temporal logic
My initial experiment to specify the "immediately committed" property
was unsuccessful. I think that I have a sense of how to write the
property using temporal logic, but I ran into snags with getting TLC
to check the property when using a state constraint in the model. I
don't have a clear enough grasp of temporal logic and how it interacts
with state constraints to fully debug this problem now. I would like to
come back to it but I feel that it might be too much of a rabbit hole
(temporal logic considered confusing?). I think a next step will be to
just augment the existing definition of immediately committed to utilize
an extra history variable so it can capture the property
fully/correctly.