Add a variant of AdvanceCommitPoint that directly inspects the global set of committed entries to advance commit point
This variant allows us to have a way to separate the correctness of
commit point advancement on primaries from commit point propagation to
secondaries. That is, we don't have to rely on the correctness of how we
calculate and compute commit point on primary in order to test if rules
of commit point propagation are safe.